requirements.txt
asgiref==3.4.1
boto==2.49.0
certifi==2021.5.30
cloudinary==1.26.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.5
django-bootstrap3==15.0.0
django-bootstrap4==0.0.6
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-heroku==0.3.1
django-herokuapp==0.9.20
django-on-heroku==1.1.2
django-registration==3.2
django-require==1.0.11
django-require-s3==1.0.0
django-storages==1.11.1
django-tinymce==3.3.0
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
gunicorn==20.1.0
importlib-metadata==2.1.1
numpy==1.19.5
opencv-python==4.2.0.32
Pillow==8.2.0
psycopg2

psycopg2-binary==2.8
pycodestyle==2.7.0
pyperclip==1.8.2
python-decouple==3.1
pytz==2020.4
sh==1.14.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.2
waitress==2.0.0
whitenoise==5.2.0
zipp==3.4.0
